Adding skipappinstall

+++ VAMobile/e2e/tests/utils.ts
@@ -568,7 +568,7 @@ export async function enableAF(AFFeature, AFUseCase, AFAppUpdate = false) {
  * @param AFUseCaseName: Name of the AF type.
  * @param AFAppUpdate: Boolean value that tells the script whether to enable the update now button or not
  * */
-export async function disableAF(featureNavigationArray, AFFeature, AFFeatureName, AFUseCaseName) {
+export async function disableAF(featureNavigationArray, AFFeature, AFFeatureName, AFUseCaseName, skipAppInstall) {
   if (AFUseCaseName === 'AllowFunction') {
     await element(by.id(CommonE2eIdConstants.HOME_TAB_BUTTON_ID)).tap()
   } else {
@@ -598,8 +598,10 @@ export async function disableAF(featureNavigationArray, AFFeature, AFFeatureName
     await expect(element(by.text(CommonE2eIdConstants.AF_ERROR_MSG_TITLE_ENTERED_TEXT))).not.toExist()
     await expect(element(by.text(CommonE2eIdConstants.AF_BODY_ENTERED_TEXT))).not.toExist()
   }
-  await device.uninstallApp()
-  await device.installApp()
+  if (!skipAppInstall) {
+    await device.uninstallApp()
+    await device.installApp()
+  }
 }
 
 /** Function that allows the AF script to navigate to a certain feature
@@ -731,7 +733,7 @@ export async function verifyAF(featureNavigationArray, AFUseCase, AFUseCaseUpgra
 
   if (AFUseCase !== 'AllowFunction') {
     if (AFUseCase === 'DenyContent' && AFUseCaseUpgrade) {
-      await disableAF(featureNavigationArray, featureNavigationArray[1], featureName, AFUseCase)
+      await disableAF(featureNavigationArray, featureNavigationArray[1], featureName, AFUseCase, undefined)
     }
   }
 }
